Cleanup image_bytes debacle
image_bytes was renamed to display_image_bytes, but this was incorrect because
it doesn't actually display anything. So undeprecate image_bytes and make
display_image_bytes actually display the image. This is a backwards
incompatible change.

Also, make image_bytes return bytes in Python 3, as that seems more
reasonable.